feat(dependecies): put tighter restrictions on peerDependencies
BREAKING CHANGE:
This commits restricts the peerDependencies from `>=` to `^`, in
order to enable better control of styles within packages depending
on this shared configuration. This package is currently in a
somewhat experimental phase, and restrictions may be opened up or
restricted further as we find what fits best for internal use at
Lime.